remove check for full versioning
This check was introduced in 42ae30c, but we are unable to figure
out why. If there was a rationale behind this, the best way to
figure it out is to revert the change and see if anything wrong happens.
This checked prevented the use of UnitOfWork::check* methods with simple
versioning.